pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astroflow"
version = "0.1.0"
description = "Reproducible analysis workflows for cosmological simulations"
readme = "README.md"
requires-python = ">=3.7"
license = "BSD-3-Clause"
license-files = ["LICENSE"]
authors = [
    {name = "Pablo Granizo Cuadrado", email = "pablogranizopgc@gmail.com"}
]
keywords = ["astrophysics", "cosmology", "simulations", "analysis", "yt", "galaxy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Framework :: Matplotlib",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Visualization"
]

dependencies = [
    "yt>=4.0.0",
    "numpy>=1.15.0",
    "matplotlib>=3.1.0",
    "pyyaml>=6.0",
    "h5py>=3.1.0",
    "unyt>=2.5.0",
    "ipykernel>=6.16.2",
    "jupyter>=1.1.1",
    "pydantic>=2.5.3",
]

[project.urls]
Homepage = "https://github.com/Ghippax/Astroflow"
Documentation = "https://astroflow.readthedocs.io"
Repository = "https://github.com/Ghippax/Astroflow"
Issues = "https://github.com/Ghippax/Astroflow/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "black>=12.0",
    "ruff>=0.6.0",
    "mypy>=1.0",
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.2",
]
analysis = [
    "scipy>=1.4.0",
]
halo = [
    "yt_astro_analysis>=1.1.0",
]

[tool.ruff]
exclude = [
    "docs",
    "*/api.py",
    "*/__init__.py",
    "*/__config__.py",
    "*.ipynb"
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=astroflow --cov-report=html --cov-report=term"
